requirements.txt
cmsplugin-filer==1.0.0
cssselect==0.9.1
Django==1.8.6
django-appconf==1.0.1
django-classy-tags==0.6.2
django-cms==3.1.3
django-filer==1.0.0
django-mptt==0.7.4
django-polymorphic==0.7.2
django-sekizai==0.8.2
Django-Select2==4.3.2
django-treebeard==3.0
djangocms-admin-style==1.0.4
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.2.0
djangocms-googlemap==0.3
djangocms-grid==1.2
djangocms-inherit==0.1
djangocms-link==1.7.1
djangocms-oembed==0.5
djangocms-picture==0.1
djangocms-snippet==1.5
djangocms-style==1.5
djangocms-table==1.2
djangocms-teaser==0.1
djangocms-text-ckeditor==2.7.0
djangocms-video==0.1
djangorestframework==3.3.1
easy-thumbnails==2.2
html5lib==0.9999999
lxml==3.4.4
micawber==0.3.3
mysqlclient==1.3.7
Pillow==3.0.0
pip==7.1.2
pyquery==1.2.9
setuptools==18.5
six==1.10.0
Unidecode==0.4.18
wheel==0.24.0